Fix constant vertical scroll bar regardless of zoom level with certain custom fonts
With some custom browser fonts, the text in div.instructions wraps more lines than intended, causing the page to be slightly taller regardless of zoom level. By offsetting .scroll-this-way by 40px from the bottom (the same amount as the top margin of .instructions), we should be able to prevent this issue.